# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/google": {
      "post": {
        "operationId": "AuthController_completeGoogle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthGoogleBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthSessionResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Complete Google sign-in with a native OAuth access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/google/oauth": {
      "post": {
        "operationId": "AuthController_completeGoogleOauth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthGoogleOauthBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthSessionResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Complete Google sign-in via authorization-code + PKCE (web)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/apple": {
      "post": {
        "operationId": "AuthController_completeApple",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthAppleBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthSessionResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Complete Apple sign-in with a native identity token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/apple/oauth": {
      "post": {
        "operationId": "AuthController_completeAppleOauth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthAppleOauthBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthSessionResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Complete Apple sign-in via authorization-code + PKCE",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/apple/oauth/session": {
      "post": {
        "operationId": "AuthController_startAppleOauthSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthAppleOauthSessionBodyDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Store Apple web OAuth PKCE session cookie before form_post",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/oauth/{provider}": {
      "post": {
        "operationId": "AuthController_completeOAuth",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "apple",
                "facebook"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Complete OAuth sign-in (Facebook, or legacy Google/Apple code/idToken path)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthSessionResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Rotate the refresh token and issue a new access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/link": {
      "post": {
        "operationId": "AuthController_linkAccount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upgrade a guest account by linking an OAuth provider",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the current session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reset-app-data": {
      "post": {
        "description": "Clears synced app data and user media while preserving the account, OAuth identity, push tokens, and sessions. Body confirmation must be the literal \"RESET\".",
        "operationId": "AuthController_resetAppData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetAppDataDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "App data reset; account and session preserved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reset app data without deleting the account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/delete-account": {
      "post": {
        "description": "Irreversible closure: wipes app data, revokes sessions, clears OAuth identity, and tombstones the email. Body must include confirmation set to the literal \"DELETE\", a primaryReason code, and optional details (max 500 characters).",
        "operationId": "AuthController_deleteAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Account closed and data removed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close account and remove user data",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Munib Tracker API",
    "description": "Cloud sync, authentication, and backend services for Munib Tracker. Track Your Journey Back to Allah.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AuthSessionResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Short-lived signed JWT access token"
          },
          "accessTokenExpiresIn": {
            "type": "number",
            "description": "Seconds until the access token expires"
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque refresh token; rotates on every refresh"
          },
          "accountType": {
            "type": "string",
            "enum": [
              "guest",
              "user"
            ]
          },
          "userId": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "google",
              "apple",
              "facebook"
            ]
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresIn",
          "refreshToken",
          "accountType",
          "userId"
        ]
      },
      "AuthGoogleBodyDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Google OAuth access token from on-device PKCE exchange"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "accountType": {
            "type": "string",
            "enum": [
              "guest",
              "user"
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "google",
              "apple",
              "facebook"
            ]
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "accountType"
        ]
      },
      "WebAuthSessionResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "AuthGoogleOauthBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "redirectUri": {
            "type": "string"
          },
          "codeVerifier": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "redirectUri",
          "codeVerifier"
        ]
      },
      "AuthAppleBodyDto": {
        "type": "object",
        "properties": {
          "identityToken": {
            "type": "string",
            "description": "Apple identity token (JWT) from native Sign in with Apple"
          },
          "displayName": {
            "type": "string",
            "description": "Given + family name from Apple (first authorization only)"
          }
        },
        "required": [
          "identityToken"
        ]
      },
      "AuthAppleOauthBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "redirectUri": {
            "type": "string"
          },
          "codeVerifier": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "redirectUri",
          "codeVerifier"
        ]
      },
      "AuthAppleOauthSessionBodyDto": {
        "type": "object",
        "properties": {
          "codeVerifier": {
            "type": "string"
          },
          "redirectUri": {
            "type": "string"
          },
          "returnUrl": {
            "type": "string",
            "description": "Web origin URL to redirect after Apple form_post callback"
          }
        },
        "required": [
          "codeVerifier",
          "redirectUri",
          "returnUrl"
        ]
      },
      "ResetAppDataDto": {
        "type": "object",
        "properties": {
          "confirmation": {
            "type": "string",
            "description": "Must be the literal string \"RESET\" to confirm the destructive data reset",
            "enum": [
              "RESET"
            ],
            "example": "RESET"
          }
        },
        "required": [
          "confirmation"
        ]
      },
      "DeleteAccountDto": {
        "type": "object",
        "properties": {
          "confirmation": {
            "type": "string",
            "description": "Must be the literal string \"DELETE\" to confirm irreversible closure",
            "enum": [
              "DELETE"
            ],
            "example": "DELETE"
          },
          "primaryReason": {
            "type": "string",
            "enum": [
              "privacy_concerns",
              "missing_features",
              "too_complicated",
              "switching_app",
              "not_using",
              "technical_issues",
              "other"
            ]
          },
          "details": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "confirmation",
          "primaryReason"
        ]
      },
...
    }
  }
}